+++ tinygrad_repo/tinygrad/shape/__init__.py
@@ -1,0 +1,4 @@
+from tinygrad.shape.shapetracker import ShapeTracker
+from tinygrad.shape.view import View
+
+__all__ = ["ShapeTracker", "View"]